scraper/requirements.txt
anyio==3.6.2
appnope==0.1.3
argon2-cffi==21.3.0
argon2-cffi-bindings==21.2.0
asttokens==2.1.0
attrs==24.2.0
backcall==0.2.0
beautifulsoup4==4.11.1
bleach==5.0.1
certifi==2024.8.30
cffi==1.15.1
charset-normalizer==3.3.2
click==8.1.7
cryptography==43.0.1
cycler==0.10.0
debugpy==1.6.3
decorator==5.1.1
defusedxml==0.7.1
dnspython==2.6.1
entrypoints==0.4
et-xmlfile==1.1.0
exceptiongroup==1.2.2
executing==1.2.0
fastjsonschema==2.16.2
filelock==3.13.4
gdown==5.1.0
greenlet==3.0.3
h11==0.14.0
idna==3.4
importlib-metadata==5.0.0
ipykernel==6.17.1
ipython==8.6.0
ipython-genutils==0.2.0
ipywidgets==8.0.2
jedi==0.18.1
Jinja2==3.1.2
joblib==1.4.0
jsonschema==4.17.0
jupyter==1.0.0
jupyter-console==6.4.4
jupyter-server==1.23.1
jupyter_client==7.4.5
jupyter_core==5.0.0
jupyterlab-pygments==0.2.2
jupyterlab-widgets==3.0.3
kiwisolver==1.3.1
MarkupSafe==2.1.1
matplotlib==3.4.2
matplotlib-inline==0.1.6
mistune==2.0.4
mysql-connector-python==8.3.0
nbclassic==0.4.8
nbclient==0.7.0
nbconvert==7.2.4
nbformat==5.7.0
nest-asyncio==1.5.6
nltk==3.8.1
notebook==6.5.2
notebook_shim==0.2.2
numpy==1.22.4
openpyxl==3.1.5
outcome==1.3.0.post0
packaging==21.3
pandas==2.2.2
pandocfilters==1.5.0
parso==0.8.3
pdfminer.six==20231228
pdfplumber==0.11.4
pexpect==4.8.0
pickleshare==0.7.5
pillow==10.4.0
platformdirs==2.5.4
prometheus-client==0.15.0
prompt-toolkit==3.0.32
psutil==5.9.4
ptyprocess==0.7.0
pure-eval==0.2.2
pycparser==2.21
Pygments==2.13.0
pymongo==4.9.1
PyMySQL==1.1.0
pyparsing==2.4.7
pypdfium2==4.30.0
pyrsistent==0.19.2
PySocks==1.7.1
pytesseract==0.3.13
python-dateutil==2.8.2
python-dotenv==1.0.1
pytz==2021.1
pyzmq==24.0.1
qtconsole==5.4.0
QtPy==2.3.0
regex==2024.4.16
requests==2.31.0
scikit-learn==1.4.2
scipy==1.7.1
seaborn==0.11.1
selenium==4.24.0
Send2Trash==1.8.0
six==1.16.0
sniffio==1.3.0
sortedcontainers==2.4.0
soupsieve==2.3.2.post1
SQLAlchemy==2.0.29
stack-data==0.6.1
terminado==0.17.0
threadpoolctl==3.4.0
tinycss2==1.2.1
tornado==6.2
tqdm==4.66.2
traitlets==5.5.0
trio==0.26.2
trio-websocket==0.11.1
typing_extensions==4.11.0
tzdata==2024.1
urllib3==2.2.1
wcwidth==0.2.5
webencodings==0.5.1
websocket-client==1.8.0
widgetsnbextension==4.0.3
wsproto==1.2.0
zipp==3.10.0
